#80[1h 30m]. Add decimals to 'records', 'accounts', 'transfers' tables. Add decimals to Record, Account, Transfer models.

+++ app/src/main/java/com/blogspot/e_kanivets/moneytracker/activity/record/MainActivity.java
@@ -228,7 +228,7 @@ public class MainActivity extends BaseDrawerActivity {
         if (defaultAccount == null) return;
 
         tvDefaultAccountTitle.setText(defaultAccount.getTitle());
-        tvDefaultAccountSum.setText(Integer.toString(defaultAccount.getCurSum()));
+        tvDefaultAccountSum.setText(Double.toString(defaultAccount.getCurSum()));
         tvCurrency.setText(defaultAccount.getCurrency());
     }
 }
